Add validation constraints to model classes
Added NotNull constraint to the fields of various model classes to ensure that they are not left blank. Moreover, a Size constraint was implemented on the 'currency' field in the ExchangeRate class to accept only three-character values. A new Jakarta Validation dependency was also included in the pom.xml file to support these modifications.