Update test parameters in ExchangeRateServiceTest
The test parameters in the ExchangeRateServiceTest have been revised. A RestTemplate mock is added and a lenient expectation is now set for it to give a mocked XML response for the getForObject method. The rate values in getLatestRates and testGetHistoricalRatesByCurrency tests are also updated. Changes align tests with recent application functionality.